finished modal card
